fix for incorrect primary key on tdata_records_* tables

+++ src/server/tools/nxdbmgr/upgrade.cpp
@@ -311,7 +311,7 @@ static BOOL CreateEventTemplate(int code, const TCHAR *name, int severity, int f
 /**
  * Re-create TDATA tables
  */
-static BOOL RecreateTData(const TCHAR *className, bool multipleTables)
+static BOOL RecreateTData(const TCHAR *className, bool multipleTables, bool indexFix)
 {
    TCHAR query[1024];
    _sntprintf(query, 256, _T("SELECT id FROM %s"), className);
@@ -321,31 +321,59 @@ static BOOL RecreateTData(const TCHAR *className, bool multipleTables)
       int count = DBGetNumRows(hResult);
       for(int i = 0; i < count; i++)
       {
+         bool recreateTables = true;
          DWORD id = DBGetFieldULong(hResult, i, 0);
 
-         if (multipleTables)
+         if (indexFix)
          {
-            _sntprintf(query, 1024, _T("DROP TABLE tdata_rows_%d\nDROP TABLE tdata_records_%d\nDROP TABLE tdata_%d\n<END>"), id, id, id);
-         }
-         else
-         {
-            _sntprintf(query, 256, _T("DROP TABLE tdata_%d\n<END>"), id);
-         }
-         if (!SQLBatch(query))
-         {
-            if (!g_bIgnoreErrors)
+            _sntprintf(query, 256, _T("SELECT count(*) FROM dc_tables WHERE node_id=%d"), id);
+            DB_RESULT hResultCount = SQLSelect(query);
+            if (hResultCount != NULL)
             {
-               DBFreeResult(hResult);
-               return FALSE;
+               recreateTables = (DBGetFieldLong(hResult, 0, 0) == 0);
+               DBFreeResult(hResultCount);
+            }
+
+            if (!recreateTables)
+            {
+               _sntprintf(query, 256, _T("CREATE INDEX idx_tdata_recid_%d ON tdata_records_%d(record_id)"), id, id);
+               if (!SQLQuery(query))
+               {
+                  if (!g_bIgnoreErrors)
+                  {
+                     DBFreeResult(hResult);
+                     return FALSE;
+                  }
+               }
             }
          }
 
-         if (!CreateTDataTables(id))
+         if (recreateTables)
          {
-            if (!g_bIgnoreErrors)
+            if (multipleTables)
             {
-               DBFreeResult(hResult);
-               return FALSE;
+               _sntprintf(query, 1024, _T("DROP TABLE tdata_rows_%d\nDROP TABLE tdata_records_%d\nDROP TABLE tdata_%d\n<END>"), id, id, id);
+            }
+            else
+            {
+               _sntprintf(query, 256, _T("DROP TABLE tdata_%d\n<END>"), id);
+            }
+            if (!SQLBatch(query))
+            {
+               if (!g_bIgnoreErrors)
+               {
+                  DBFreeResult(hResult);
+                  return FALSE;
+               }
+            }
+
+            if (!CreateTDataTables(id))
+            {
+               if (!g_bIgnoreErrors)
+               {
+                  DBFreeResult(hResult);
+                  return FALSE;
+               }
             }
          }
       }
@@ -356,6 +384,29 @@ static BOOL RecreateTData(const TCHAR *className, bool multipleTables)
       if (!g_bIgnoreErrors)
          return FALSE;
    }
+   return TRUE;
+}
+
+/**
+ * Upgrade from V323 to V324
+ */
+static BOOL H_UpgradeFromV323(int currVersion, int newVersion)
+{
+   if (!MetaDataReadInt(_T("ValidTDataPK"), 0))  // check if schema is already correct
+   {
+      TCHAR query[1024];
+      _sntprintf(query, 1024, 
+         _T("UPDATE metadata SET var_value='CREATE TABLE tdata_records_%%d (record_id %s not null,row_id %s not null,instance varchar(255) null,PRIMARY KEY(record_id),FOREIGN KEY (record_id) REFERENCES tdata_%%d(record_id) ON DELETE CASCADE)' WHERE var_name='TDataTableCreationCommand_1'"),
+         g_pszSqlType[g_iSyntax][SQL_TYPE_INT64], g_pszSqlType[g_iSyntax][SQL_TYPE_INT64]);
+      CHK_EXEC(SQLQuery(query));
+
+      RecreateTData(_T("nodes"), true, true);
+      RecreateTData(_T("clusters"), true, true);
+      RecreateTData(_T("mobile_devices"), true, true);
+   }
+
+   CHK_EXEC(SQLQuery(_T("DELETE FROM metadata WHERE var_name='ValidTDataPK'")));
+   CHK_EXEC(SQLQuery(_T("UPDATE metadata SET var_value='324' WHERE var_name='SchemaVersion'")));
    return TRUE;
 }
 
@@ -967,16 +1018,17 @@ static BOOL H_UpgradeFromV293(int currVersion, int newVersion)
       _T("INSERT INTO metadata (var_name,var_value)")
       _T("   VALUES ('TDataTableCreationCommand_0','CREATE TABLE tdata_%d (item_id integer not null,tdata_timestamp integer not null,record_id $SQL:INT64 not null,UNIQUE(record_id))')\n")
       _T("INSERT INTO metadata (var_name,var_value)")
-	   _T("   VALUES ('TDataTableCreationCommand_1','CREATE TABLE tdata_records_%d (record_id $SQL:INT64 not null,row_id $SQL:INT64 not null,instance varchar(255) null,PRIMARY KEY(row_id),FOREIGN KEY (record_id) REFERENCES tdata_%d(record_id) ON DELETE CASCADE)')\n")
+	   _T("   VALUES ('TDataTableCreationCommand_1','CREATE TABLE tdata_records_%d (record_id $SQL:INT64 not null,row_id $SQL:INT64 not null,instance varchar(255) null,PRIMARY KEY(record_id),FOREIGN KEY (record_id) REFERENCES tdata_%d(record_id) ON DELETE CASCADE)')\n")
       _T("INSERT INTO metadata (var_name,var_value)")
 	   _T("   VALUES ('TDataTableCreationCommand_2','CREATE TABLE tdata_rows_%d (row_id $SQL:INT64 not null,column_id integer not null,value varchar(255) null,PRIMARY KEY(row_id,column_id),FOREIGN KEY (row_id) REFERENCES tdata_records_%d(row_id) ON DELETE CASCADE)')\n")
       _T("<END>");
    CHK_EXEC(SQLBatch(batch));
 
-   RecreateTData(_T("nodes"), true);
-   RecreateTData(_T("clusters"), true);
-   RecreateTData(_T("mobile_devices"), true);
+   RecreateTData(_T("nodes"), true, false);
+   RecreateTData(_T("clusters"), true, false);
+   RecreateTData(_T("mobile_devices"), true, false);
 
+   CHK_EXEC(SQLQuery(_T("INSERT INTO metadata (var_name,var_value) VALUES ('ValidTDataPK','1')")));
    CHK_EXEC(SQLQuery(_T("UPDATE metadata SET var_value='294' WHERE var_name='SchemaVersion'")));
    return TRUE;
 }
@@ -1191,9 +1243,9 @@ static BOOL H_UpgradeFromV280(int currVersion, int newVersion)
       _T("<END>");
    CHK_EXEC(SQLBatch(batch));
 
-   RecreateTData(_T("nodes"), false);
-   RecreateTData(_T("clusters"), false);
-   RecreateTData(_T("mobile_devices"), false);
+   RecreateTData(_T("nodes"), false, false);
+   RecreateTData(_T("clusters"), false, false);
+   RecreateTData(_T("mobile_devices"), false, false);
 
    CHK_EXEC(SQLQuery(_T("UPDATE metadata SET var_value='281' WHERE var_name='SchemaVersion'")));
    return TRUE;
@@ -7834,6 +7886,7 @@ static struct
    { 320, 321, H_UpgradeFromV320 },
    { 321, 322, H_UpgradeFromV321 },
    { 322, 323, H_UpgradeFromV322 },
+   { 323, 324, H_UpgradeFromV323 },
    { 0, 0, NULL }
 };
